+++ rebench/rebench.py
@@ -184,7 +184,8 @@ Argument:
         executor = Executor(runs, self._config.use_nice,
                             self._config.options.include_faulty,
                             self._config.options.verbose,
-                            scheduler_class)
+                            scheduler_class,
+                            self._config.build_log)
         return executor.execute()