+++ rebench/tests/features/issue_58_buildvm_a.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+echo "#!/bin/bash" > vm_58a.sh
+echo "echo \$@" >> vm_58a.sh
+chmod +x vm_58a.sh
+echo standard
+echo error 1>&2